Adjusts content sizes according to the chosen Dynamic Type size
